simple/requirements.txt
python-osc==1.9.3